// client/src/redux/auth/auth.types.ts
export enum AuthTypes {
  SIGN_IN_START = "SIGN_IN_START",
  SIGN_IN_SUCCESS = "SIGN_IN_SUCCESS",
  SIGN_IN_FAILURE = "SIGN_IN_FAILURE",
  SIGN_UP_START = "SIGN_UP_START",
  SIGN_UP_SUCCESS = "SIGN_UP_SUCCESS",
  SIGN_UP_FAILURE = "SIGN_UP_FAILURE",
  SIGN_OUT_START = "SIGN_OUT_START",
  SIGN_OUT_SUCCESS = "SIGN_OUT_SUCCESS",
  SIGN_OUT_FAILURE = "SIGN_OUT_FAILURE",
  RESET_SIGN_UP_ERRORS = "RESET_SIGN_UP_ERRORS",
  GET_ACCESS_TOKEN_START = "GET_ACCESS_TOKEN_START",
  GET_ACCESS_TOKEN_SUCCESS = "GET_ACCESS_TOKEN_SUCCESS",
  GET_ACCESS_TOKEN_FAILURE = "GET_ACCESS_TOKEN_FAILURE",
  GET_CURRENT_USER_START = "GET_CURRENT_USER_START",
  GET_CURRENT_USER_SUCCESS = "GET_CURRENT_USER_SUCCESS",
  GET_CURRENT_USER_FAILURE = "GET_CURRENT_USER_FAILURE",
}